Make the cache synchronous and switch to `DashMap` in its internals (#1444)
This commit aims to make the cache not require an asynchronous context to be used, along with removing locks for a way lower probability of a deadlock or poisoned locks, this was done by replacing `tokio::RwLock<HashMap>` with `DashMap` and `parking_lot::RwLock<...>` for other types that are neither Map nor Set.

Using `parking_lot` does not pull an extra dependency, as it's already a dependency of `tokio`. `DashMap` is the only new dependency, which itself doesn't depend on much.

This is a step forward to a library that depends less on `tokio`, so most of it features can compile to wasm, so they can be used with [wasm_plugin](https://github.com/alec-deason/wasm_plugin/) to allow for dynamic runtime functions and commands in the future.